requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
boto3==1.40.26
botocore==1.40.26
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-storages==1.14.6
idna==3.10
jmespath==1.0.1
lxml==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0